Move composite back out of library
